+++ CHANGELOG.md
@@ -9,6 +9,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 * `equal_with_nan` in `util_numpy`
 
+### Fixed:
+* Fixed issue in RunningStats where computation was incorrect when weights were given.
+* RunningStats now handles nans.
+
 ## Version 0.6.2 - Released 2022-06-15
 
 ### Added